helperscripts/requirements.txt
psycopg2-binary>=2.9.0
tqdm>=4.60.0